backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
certifi
langchain-groq
langchain
langchain-core
# google-generativeai removed
bcrypt
PyJWT
waitress
crewai
litellm
